removed unnecessary casting

+++ app/src/main/java/com/laudien/p1xelfehler/batterywarner/HelperClasses/NotificationHelper.java
@@ -364,7 +364,7 @@ public final class NotificationHelper {
                 RemoteViews contentView = new RemoteViews(context.getPackageName(), R.layout.notification_battery_info);
                 contentView.setImageViewResource(R.id.img_battery, R.mipmap.ic_launcher);
                 int textViewId;
-                for (byte i = 0; i < data.length; i++) {
+                for (int i = 0; i < data.length; i++) {
                     textViewId = BatteryHelper.getTextViewId(i);
                     if (data[i] != null) {
                         contentView.setTextViewText(textViewId, data[i]);
